Add files via upload
4000 Customer Submits Application
Most automation systems start the functionality with the applicant/client/customer submitting their case/application.  There is a typical range of information that is included in an application addressed in the 4000 level of the case management framework.  An issue is how much verification is done at the time of submittal versus later in the review process.  There should be enough verification at the time of submittal to avoid doing work on the application only to find out the applicant is not eligible; but the application should not be so difficult to submit that legitimate candidates are too frustrated.