7000 Verify Application Cross Reference
The 7000 Verify Application Cross Reference performs the detailed verification of the applicant and the application supporting documentation.  Some verification tasks are performed at the time the application is submitted.  But more detailed verification typically occurs later due to the time and expense required to access third-party data sources.  This can include any history of the applicant's past performance, associations, and more