9000 Inspection Complaint Appeal
The 9000 Theme addresses a wide range of activity that ensure compliance with the conditions of the original approval or allow appeal of the denial or citation.  The range of activities include:  public complaint processing and inhouse resolution, recipient reporting to verify adherence to terms and field inspections, Investigations where complaints or recipient reporting requires further effort, citations for enforcement, and appeals.  Appeals may be for denial of an application or response to a citation.