3000 Account Management
Applicants may be individuals or organizations.  Online submittal of an application made possible by enabling the applicant to establish an account that involves authenticating identity.  Once the account is established the representative can submit an application, amendments, respond to inquiries, authorize third party reference checks, and provide electronic signature for all official actions related to the application and its approval.  Accounts may have multiple participants each with a different level of authorized activity on behalf of the account.